+++ changelog.md
@@ -1,10 +1,5 @@
 # Changelog
 
-#### Segment class:
-- properties:
-    - lenght
-    - square_lenght
-
 #### Point class
 - methods:
     - distance
@@ -16,9 +11,29 @@
 #### Vector class:
 - properties:
     - square_lenght
+    - magnitude
 - methods:
     - angle
     - project_on
+    - copy
+
+#### Segment class:
+- properties:
+    - lenght
+    - square_lenght
+- methods:
+    - translate
+    - copy
+
+#### Ray class:
+- methods:
+    - translate
+    - copy
+
+#### Line class:
+- methods:
+    - translate
+    - copy
 
 #### Circle class _(new)_:
 - properties:
@@ -27,11 +42,11 @@
     - intersect
     - count_intersections
     - get_intersection
+    - copy
     - \_\_contains__
     - \_\_eq__
 
 ## Comming soon
-- translate method on Vector, Segment, Line
-- rotate method on Vector, Segment, Line
+- rotate method on Segment, Line
 - Rectangle, Polygon, Ellipse
 - all maths demos